Refactor the demo app to make it easy to reuse the common part across the playground that uses FlexboxLayout and RecyclerView.LayoutManager.
To make it easy to reuse the common part across the playground that uses FlexboxLayout and RecyclerView.LayoutManager by extracting the common logic from the MainActivity into other classes.

Follow up CL is in Progress to add a Fragment for RecyclerView.LayoutManager.
