+++ app/src/main/java/com/google/android/apps/flexbox/FlexItemEditFragment.java
@@ -16,6 +16,7 @@
 
 package com.google.android.apps.flexbox;
 
+import com.google.android.apps.flexbox.listeners.FlexItemChangedListener;
 import com.google.android.apps.flexbox.validators.DimensionInputValidator;
 import com.google.android.apps.flexbox.validators.FixedDimensionInputValidator;
 import com.google.android.apps.flexbox.validators.FlexBasisPercentInputValidator;
@@ -453,13 +454,5 @@ public class FlexItemEditFragment extends DialogFragment {
                     break;
             }
         }
-    }
-
-    /**
-     * A listener that listens to the change of a flex item
-     */
-    public interface FlexItemChangedListener {
-
-        void onFlexItemChanged(FlexItem flexItem, int viewIndex);
     }
 }
