+++ app/src/main/java/com/google/android/apps/flexbox/listeners/FlexItemChangedListener.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2016 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.apps.flexbox.listeners;
+
+import com.google.android.flexbox.FlexItem;
+
+/**
+ * A listener that listens to the change of a flex item
+ */
+public interface FlexItemChangedListener {
+
+    void onFlexItemChanged(FlexItem flexItem, int viewIndex);
+}
