+++ app/src/main/res/layout/content_main.xml
@@ -13,37 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
-<com.google.android.flexbox.FlexboxLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/flexbox_layout"
-    android:layout_width="match_parent"
+    android:id="@+id/container"
     android:layout_height="match_parent"
-    app:alignContent="flex_start"
-    app:alignItems="flex_start"
-    app:flexWrap="wrap"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:showIn="@layout/activity_main">
+    android:layout_width="match_parent"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
-    <TextView
-        android:id="@+id/textview1"
-        style="@style/FlexItem"
-        android:layout_width="@dimen/flex_item_length2"
-        android:layout_height="@dimen/flex_item_length"
-        android:text="@string/one" />
-
-    <TextView
-        android:id="@+id/textview2"
-        style="@style/FlexItem"
-        android:layout_width="@dimen/flex_item_length3"
-        android:layout_height="@dimen/flex_item_length"
-        android:text="@string/two" />
-
-    <TextView
-        android:id="@+id/textview3"
-        style="@style/FlexItem"
-        android:layout_width="@dimen/flex_item_length"
-        android:layout_height="@dimen/flex_item_length"
-        android:text="@string/three" />
-
-</com.google.android.flexbox.FlexboxLayout>
