+++ app/src/main/res/layout/fragment_flexboxlayout.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+Copyright 2016 Google Inc. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+<com.google.android.flexbox.FlexboxLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/flexbox_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:alignContent="flex_start"
+    app:alignItems="flex_start"
+    app:flexWrap="wrap"
+    tools:showIn="@layout/activity_main">
+
+    <TextView
+        android:id="@+id/textview1"
+        style="@style/FlexItem"
+        android:layout_width="@dimen/flex_item_length2"
+        android:layout_height="@dimen/flex_item_length"
+        android:text="@string/one" />
+
+    <TextView
+        android:id="@+id/textview2"
+        style="@style/FlexItem"
+        android:layout_width="@dimen/flex_item_length3"
+        android:layout_height="@dimen/flex_item_length"
+        android:text="@string/two" />
+
+    <TextView
+        android:id="@+id/textview3"
+        style="@style/FlexItem"
+        android:layout_width="@dimen/flex_item_length"
+        android:layout_height="@dimen/flex_item_length"
+        android:text="@string/three" />
+
+</com.google.android.flexbox.FlexboxLayout>
